requirements.txt
requests==2.31.0
AnilistPython==0.1.3
python-telegram-bot==20.3